add defcustom for ag-path
On windows, "ag" doesn't work for me, only "ag.exe". Additionally, I
am using a package manager
((scoop)[https://github.com/lukesampson/scoop]) that puts the actual
executable in a version dependent folder and a wrapper script in the path.

All this leads me to the conclusion that the simplest thing is just to
have a setting for the ag executable to use.